Text.FuzzyFind

Description

A package that provides an API for fuzzy text search in Haskell, using a modified version of the Smith-Waterman algorithm. The search is intended to behave similarly to the excellent fzf tool by Junegunn Choi.

Documentation

bestMatch

Arguments

:: String
The query pattern.

-> String
The input string.

-> Maybe Alignment

bestMatch query string will return Nothing if query is not a subsequence of string. Otherwise, it will return the "best" way to line up the characters in query with the characters in string. Lower-case characters in the query are assumed to be case-insensitive, and upper-case characters are assumed to be case-sensitive.

For example:

> bestMatch "ff" "FuzzyFind"
Just (Alignment {score = 25, result = Result {[Match "F", Gap "uzzy", Match "F", Gap "ind"]}})

The score indicates how "good" the match is. Better matches have higher scores. There's no maximum score (except for the upper limit of the Int datatype), but the lowest score is 0.

A substring from the query will generate a Match, and any characters from the input that don't result in a Match will generate a Gap. Concatenating all the Match and Gap results should yield the original input string.

Note that the matched characters in the input always occur in the same order as they do in the query pattern.

The algorithm prefers (and will generate higher scores for) the following kinds of matches:

Contiguous characters from the query string. For example, bestMatch "pp" will find the last two ps in "pickled pepper".

Characters at the beginnings of words. For example, bestMatch "pp" will find the first two Ps in "Peter Piper".

Characters at CamelCase humps. For example, bestMatch "bm" "BatMan" will score higher than bestMatch "bm" "Batman".

The algorithm strongly prefers the first character of the query pattern to be at the beginning of a word or CamelHump. For example, bestMatch "mn" "Bat Man" will score higher than bestMatch "atn" "Batman".

All else being equal, matches that occur later in the input string are preferred.

fuzzyFind

Arguments

:: [String]
The query patterns.

-> [String]
The input strings.

-> [Alignment]

Finds input strings that match all the given input patterns. For each input that matches, it returns one Alignment. The output is not sorted. ascending.

For example:

> import Data.Foldable
> traverse_ (putStrLn . ("\n" ++) . highlight) $ fuzzyFind ["dad", "mac", "dam"] ["red macadamia", "Madam Card"]

Madam Card
* *** ** *

red macadamia
  * *******

fuzzyFindOn :: (a -> String) -> [String] -> [a] -> [(Alignment, a)]

A version of fuzzyFind that searches on the given text field of the data.

type Score = Int

data Alignment

An Alignment is a Score together with a Result. Better results have higher scores.

Constructors

Alignment

Fields

score :: !Score

result :: !Result

defaultMatchScore :: Int

The base score given to a matching character

defaultMismatchScore :: Int

The base score given to a mismatched character

defaultBoundaryBonus :: Int

Bonus points given to characters matching at the beginning of words

defaultCamelCaseBonus :: Int

Bonus points given to characters matching a hump of a CamelCase word. We subtract a point from the word boundary score, since a word boundary will incur a gap penalty.

defaultFirstCharBonusMultiplier :: Int

Double any bonus points for matching the first pattern of the character. This way we strongly prefer starting the match at the beginning of a word.

defaultGapPenalty :: Int

We prefer consecutive runs of matched characters in the pattern, so we impose a penalty for any gaps, which is added to the size of the gap.

defaultConsecutiveBonus :: Int

We give a bonus to consecutive matching characters. A number about the same as the boundary bonus will prefer runs of consecutive characters vs finding acronyms.

defaultLaterBonusMultiplier :: Int

We give a bonus for matches that occur later in the input string. If this is e.g. 100, we give one bonus point for each percentage of the length of the input that the match occurs at (100 points for the last character, 50 for characters in the middle, and so on).

segmentToString :: ResultSegment -> String

highlight :: Alignment -> String

Renders an Alignment as a pair of lines with "*" on the lower line indicating the location of pattern matches.

highlight' :: Alignment -> Text

bestMatch'

Arguments

:: Int
Base score for a matching character. See defaultMatchScore.

-> Int
Base score for a mismatched character. See defaultMismatchScore.

-> Int
Additional penalty for a gap. See defaultGapPenalty.

-> Int
Bonus score for a match at the beginning of a word. See defaultBoundaryBonus.

-> Int
Bonus score for a match on a CamelCase hump. See defaultCamelCaseBonus.

-> Int
Bonus multiplier for matching the first character of the pattern. See defaultFirstCharBonusMultiplier.

-> Int
Bonus score for each consecutive character matched. See defaultFirstCharBonusMultiplier.

-> Int
Bonus multiplier for matching later characters of the input. E.g. if this is 10, the bonus for matching the last character of the input is 10, and 5 for characters close to the middle, and so on.

-> String
The query pattern.

-> String
The input string.

-> Maybe Alignment

A highly configurable version of bestMatch.

data ResultSegment

Constructors

Gap !String

Match !String

newtype Result

Concatenating all the ResultSegments should yield the original input string.

Constructors

Result

Fields

segments :: Seq ResultSegment

mergeResults :: Result -> Result -> Result